rename: directory 구조 수정
