+++ src/main/java/teamyc/recordpet/domain/user/controller/UserController.java
@@ -1,13 +1,13 @@
-package teamyc.recordpet.user.controller;
+package teamyc.recordpet.domain.user.controller;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
-import teamyc.recordpet.user.dto.UserSignupRequest;
-import teamyc.recordpet.user.dto.UserSignupResponse;
-import teamyc.recordpet.user.service.UserService;
+import teamyc.recordpet.domain.user.dto.UserSignupRequest;
+import teamyc.recordpet.domain.user.dto.UserSignupResponse;
+import teamyc.recordpet.domain.user.service.UserService;
 
 @RestController
 @RequestMapping("/api/v1/users")
