+++ src/main/java/teamyc/recordpet/domain/user/repository/UserRepository.java
@@ -1,7 +1,7 @@
-package teamyc.recordpet.user.repository;
+package teamyc.recordpet.domain.user.repository;
 
 import org.springframework.data.repository.RepositoryDefinition;
-import teamyc.recordpet.user.entity.User;
+import teamyc.recordpet.domain.user.entity.User;
 
 @RepositoryDefinition(domainClass = User.class, idClass = Long.class)
 public interface UserRepository {
